requirements.txt
pytest==7.2.1
Appium-Python-Client==2.7.1
selenium~=4.7.2
python-dotenv == 0.21.0
